Revert "Updated Trip and User"
This reverts commit dff902c3c426de4a786c962a68444aca6179111f.

+++ src/main/java/com/vroommates/VroomMates/model/tripmodel/Trip.java
@@ -21,10 +21,10 @@ public class Trip {
     private boolean IsLive;
 
     //start coordinates
-    private long StartLat;
-    private long StartLon;
+    private int StartLat;
+    private int StartLon;
 
     //end coordinates
-    private long EndLat;
-    private long EndLon;
+    private int EndLat;
+    private int EndLon;
 }
